pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcglauber"
version = "0.2.0"
description = "Modern modular Monte Carlo Glauber model with interpolated sigma_NN and event-level geometry outputs"
authors = [{name = "OpenAI"}]
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "matplotlib>=3.7",
  "pyyaml>=6.0",
  "tqdm>=4.66"
]

[project.scripts]
mcglauber = "mcglauber.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
